+++ ceammc/ext/src/flow/seq_bangs.h
@@ -31,6 +31,7 @@ public:
     void onBang() override;
     void onInlet(size_t n, const AtomList& l) override;
 
+    void m_stop(t_symbol* s, const AtomListView& lv);
     void m_reset(t_symbol* s, const AtomListView& lv);
 
 protected: