+++ ceammc/CHANGELOG.md
@@ -47,6 +47,7 @@
 - an.rms~, root mean square with moving-average algorithm
 - midi.vramp added
 - midi.tuning added
+- seq.bangs sequencer added into new category: seq
 
 
 ### Changed: